pyproject.toml
[project]
name = "dobby-benchmark-script"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch",
    "torchao",
    "diffusers",
    "pandas",
    "matplotlib",
    "transformers",
    "accelerate",
    "Pillow",
    "optimum-quanto>=0.2.7",
]


[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }
torchaudio = { index = "pytorch-cu130" }
xformers = { index = "pytorch-cu130" }
torchao = { index = "pytorch-cu130" }

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
